Moving unsubscribe monthly action to user controller and out of statistics controller
