+++ app/controllers/user_controller.rb
@@ -1,5 +1,5 @@
 class UserController < ApplicationController
-  before_action :authenticate_user!
+  before_action :authenticate_user!, except: :unsubscribe_monthly
 
   layout 'dashboard'
 
@@ -9,6 +9,25 @@ class UserController < ApplicationController
     @pending_works = pending_works
     @embargoed_works = embargoed_works
     @current_works_with_stats = current_works_with_stats
+  end
+
+  def unsubscribe_monthly
+    author_id = params[:author_id].to_s
+
+    begin
+      raise 'Request missing parameters.' if author_id.blank? || params[:chk].blank?
+      raise 'Cannot be verified.' unless Rails.application.message_verifier(:unsubscribe).verify(params[:chk]) == author_id
+
+      epref = EmailPreference.find_or_initialize_by(author: author_id)
+      epref.monthly_opt_out = true
+      epref.save!
+
+      flash[:success] = 'Unsubscribe request successful'
+    rescue StandardError
+      flash[:error] = 'There was an error with your unsubscribe request'
+    end
+
+    redirect_to root_url
   end
 
   private
